⚰️ Remove dead code
