+++ pkg/threepilestrick/trick.go
@@ -1,7 +1,10 @@
 package threepilestrick
 
 import (
+	"errors"
 	"fmt"
+
+	"github.com/carloscasalar/card-guess/internal/mat"
 
 	"github.com/carloscasalar/card-guess/internal/deck"
 )
@@ -51,4 +54,21 @@ func pileToSerializablePile(pile deck.Pile) Pile {
 		cards = append(cards, Card(card.String()))
 	}
 	return cards
+}
+
+func splitIntoThreePiles(sample deck.Pile) (*mat.Mat, error) {
+	theMat := mat.New()
+	for {
+		var card deck.Card
+		var err error
+		card, sample, err = sample.DrawCard()
+		if err != nil {
+			if errors.Is(err, deck.ErrNoMoreCardsInThePile) {
+				break
+			}
+			return nil, err
+		}
+		theMat = theMat.PlaceIntoNextPile(card)
+	}
+	return &theMat, nil
 }
